backend/pyproject.toml
[project]
name = "backend"
version = "0.0.1"
description = "Backend for wheel-n-deal, a package for tracking flash sales and prices across specific items"
authors = [{ name = "Jonathan Kirkland", email = "jbkirkland708@gmail.com" }]
license = { text = "GNU General Public License v3.0" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.31.1",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.35",
    "databases>=0.9.0",
    "asyncpg>=0.29.0",
    "selenium>=4.25.0",
    "celery>=5.4.0",
    "redis>=5.0.1",
    "pydantic-settings>=2.2.1",
    "loguru>=0.7.2",
    "prometheus-client>=0.19.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.1.2,<5.0.0",
    "python-multipart>=0.0.9",
    "waitress>=2.1.2",
    "slowapi>=0.1.9",
    "lxml>=5.1.0",
    "alembic>=1.13.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.6.9",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
    "ty>=0.0.12",
    "pre-commit>=4.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=. --cov-report=term-missing --cov-report=xml:coverage.xml"

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "**/__pycache__/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
